% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinksData.R
\docType{package}
\name{TCGAbiolinksData}
\alias{TCGAbiolinksData}
\alias{TCGAbiolinksData-package}
\title{Data for TCGAbiolinks package}
\description{
This package provides all necessary data to run and compile TCGAbiolinks package.
For more information how to create these objects please read the vignette of this package with the
follwing command: \code{browseVignettes("TCGAbiolinksData")}
}
\examples{
data("case1")
data("case2")
data("case3")
data("geneInfoHT")
data("geneInfo")
}
\keyword{utilities}
